#95 completely removed substitutions from point based heap analyses
