+++ lisa/imp-testcases/heap/point-based-heap/field-insensitive/analysis___untyped_A.f2(A_this).dot
@@ -1,14 +1,14 @@
 digraph {
-	"node0" [shape="rect",color="black",label=<o = new A(-1, -1)<BR/>{{<BR/>heap [[ [heap[s]:pp@'imp-testcases/heap/point-based-heap/field-insensitive/program.imp':13:21] ]]<BR/>value [[ #TOP# ]]<BR/>}} -&gt; [o]>];
-	"node1" [shape="rect",color="gray",label=<o = new A(0, 0)<BR/>{{<BR/>heap [[ [heap[s]:pp@'imp-testcases/heap/point-based-heap/field-insensitive/program.imp':15:16] ]]<BR/>value [[ #TOP# ]]<BR/>}} -&gt; [o]>];
-	"node2" [shape="rect",color="gray",label=<true<BR/>{{<BR/>heap [[ [heap[s]:pp@'imp-testcases/heap/point-based-heap/field-insensitive/program.imp':13:21] ]]<BR/>value [[ #TOP# ]]<BR/>}} -&gt; [true]>];
-	"node3" [shape="rect",color="black",peripheries="2",label=<ret<BR/>{{<BR/>heap [[ [heap[s]:pp@'imp-testcases/heap/point-based-heap/field-insensitive/program.imp':15:16, heap[s]:pp@'imp-testcases/heap/point-based-heap/field-insensitive/program.imp':17:16] ]]<BR/>value [[ #TOP# ]]<BR/>}} -&gt; [skip]>];
-	"node4" [shape="rect",color="gray",label=<o = new A(1, 1)<BR/>{{<BR/>heap [[ [heap[s]:pp@'imp-testcases/heap/point-based-heap/field-insensitive/program.imp':17:16] ]]<BR/>value [[ #TOP# ]]<BR/>}} -&gt; [o]>];
-	"node1" -> "node3" [color="black"];
-	"node0" -> "node2" [color="black"];
-	"node2" -> "node1" [color="blue",style="dashed"];
-	"node2" -> "node4" [color="red",style="dashed"];
-	"node4" -> "node3" [color="black"];
+	"node0" [shape="rect",color="black",label=<o = new A(-1, -1)<BR/>{{<BR/>heap [[ [heap[w]:pp@'imp-testcases/heap/point-based-heap/field-insensitive/program.imp':13:21] ]]<BR/>value [[ #TOP# ]]<BR/>}} -&gt; [o]>];
+	"node1" [shape="rect",color="gray",label=<o = new A(0, 0)<BR/>{{<BR/>heap [[ [heap[w]:pp@'imp-testcases/heap/point-based-heap/field-insensitive/program.imp':15:16] ]]<BR/>value [[ #TOP# ]]<BR/>}} -&gt; [o]>];
+	"node2" [shape="rect",color="gray",label=<o = new A(1, 1)<BR/>{{<BR/>heap [[ [heap[w]:pp@'imp-testcases/heap/point-based-heap/field-insensitive/program.imp':17:16] ]]<BR/>value [[ #TOP# ]]<BR/>}} -&gt; [o]>];
+	"node3" [shape="rect",color="gray",label=<true<BR/>{{<BR/>heap [[ [heap[w]:pp@'imp-testcases/heap/point-based-heap/field-insensitive/program.imp':13:21] ]]<BR/>value [[ #TOP# ]]<BR/>}} -&gt; [true]>];
+	"node4" [shape="rect",color="black",peripheries="2",label=<ret<BR/>{{<BR/>heap [[ [heap[w]:pp@'imp-testcases/heap/point-based-heap/field-insensitive/program.imp':15:16, heap[w]:pp@'imp-testcases/heap/point-based-heap/field-insensitive/program.imp':17:16] ]]<BR/>value [[ #TOP# ]]<BR/>}} -&gt; [skip]>];
+	"node1" -> "node4" [color="black"];
+	"node2" -> "node4" [color="black"];
+	"node3" -> "node1" [color="blue",style="dashed"];
+	"node3" -> "node2" [color="red",style="dashed"];
+	"node0" -> "node3" [color="black"];
 subgraph cluster_legend {
 	label="Legend";
 	style=dotted;
