+++ lisa/imp-testcases/heap/point-based-heap/field-insensitive/analysis___untyped_A.f5(A_this).dot
@@ -1,16 +1,16 @@
 digraph {
 	"node0" [shape="rect",color="black",label=<i = 0<BR/>{{<BR/>heap [[ #TOP# ]]<BR/>value [[ i: [0, 0] ]]<BR/>}} -&gt; [i]>];
-	"node1" [shape="rect",color="gray",label=<true<BR/>{{<BR/>heap [[ [heap[s]:pp@'imp-testcases/heap/point-based-heap/field-insensitive/program.imp':45:19, heap[w]:pp@'imp-testcases/heap/point-based-heap/field-insensitive/program.imp':47:16] ]]<BR/>value [[ heap[w]:pp@'imp-testcases/heap/point-based-heap/field-insensitive/program.imp':47:16: [0, 0]<BR/>i: [0, +Inf] ]]<BR/>}} -&gt; [true]>];
+	"node1" [shape="rect",color="gray",label=<true<BR/>{{<BR/>heap [[ [heap[w]:pp@'imp-testcases/heap/point-based-heap/field-insensitive/program.imp':45:19, heap[w]:pp@'imp-testcases/heap/point-based-heap/field-insensitive/program.imp':47:16] ]]<BR/>value [[ heap[w]:pp@'imp-testcases/heap/point-based-heap/field-insensitive/program.imp':47:16: [0, 0]<BR/>i: [0, +Inf] ]]<BR/>}} -&gt; [true]>];
 	"node2" [shape="rect",color="gray",label=<o = new int[](6)<BR/>{{<BR/>heap [[ [heap[w]:pp@'imp-testcases/heap/point-based-heap/field-insensitive/program.imp':47:16] ]]<BR/>value [[ heap[w]:pp@'imp-testcases/heap/point-based-heap/field-insensitive/program.imp':47:16: [0, 0]<BR/>i: [0, +Inf] ]]<BR/>}} -&gt; [o]>];
-	"node3" [shape="rect",color="black",peripheries="2",label=<ret<BR/>{{<BR/>heap [[ [heap[s]:pp@'imp-testcases/heap/point-based-heap/field-insensitive/program.imp':45:19, heap[w]:pp@'imp-testcases/heap/point-based-heap/field-insensitive/program.imp':47:16] ]]<BR/>value [[ heap[w]:pp@'imp-testcases/heap/point-based-heap/field-insensitive/program.imp':47:16: [0, 0]<BR/>i: [0, +Inf] ]]<BR/>}} -&gt; [skip]>];
+	"node3" [shape="rect",color="black",peripheries="2",label=<ret<BR/>{{<BR/>heap [[ [heap[w]:pp@'imp-testcases/heap/point-based-heap/field-insensitive/program.imp':45:19, heap[w]:pp@'imp-testcases/heap/point-based-heap/field-insensitive/program.imp':47:16] ]]<BR/>value [[ heap[w]:pp@'imp-testcases/heap/point-based-heap/field-insensitive/program.imp':47:16: [0, 0]<BR/>i: [0, +Inf] ]]<BR/>}} -&gt; [skip]>];
 	"node4" [shape="rect",color="gray",label=<i = +(i, 1)<BR/>{{<BR/>heap [[ [heap[w]:pp@'imp-testcases/heap/point-based-heap/field-insensitive/program.imp':47:16] ]]<BR/>value [[ heap[w]:pp@'imp-testcases/heap/point-based-heap/field-insensitive/program.imp':47:16: [0, 0]<BR/>i: [1, +Inf] ]]<BR/>}} -&gt; [i]>];
 	"node5" [shape="rect",color="gray",label=<[](o, 0) = 0<BR/>{{<BR/>heap [[ [heap[w]:pp@'imp-testcases/heap/point-based-heap/field-insensitive/program.imp':47:16] ]]<BR/>value [[ heap[w]:pp@'imp-testcases/heap/point-based-heap/field-insensitive/program.imp':47:16: [0, 0]<BR/>i: [0, +Inf] ]]<BR/>}} -&gt; [heap[w]:pp@'imp-testcases/heap/point-based-heap/field-insensitive/program.imp':47:16]>];
-	"node6" [shape="rect",color="gray",label=<o = new int[](6)<BR/>{{<BR/>heap [[ [heap[s]:pp@'imp-testcases/heap/point-based-heap/field-insensitive/program.imp':45:19] ]]<BR/>value [[ i: [0, 0] ]]<BR/>}} -&gt; [o]>];
+	"node6" [shape="rect",color="gray",label=<o = new int[](6)<BR/>{{<BR/>heap [[ [heap[w]:pp@'imp-testcases/heap/point-based-heap/field-insensitive/program.imp':45:19] ]]<BR/>value [[ i: [0, 0] ]]<BR/>}} -&gt; [o]>];
 	"node1" -> "node2" [color="blue",style="dashed"];
 	"node1" -> "node3" [color="red",style="dashed"];
 	"node2" -> "node5" [color="black"];
-	"node4" -> "node1" [color="black"];
 	"node0" -> "node6" [color="black"];
+	"node4" -> "node1" [color="black"];
 	"node5" -> "node4" [color="black"];
 	"node6" -> "node1" [color="black"];
 subgraph cluster_legend {
