+++ lisa/imp-testcases/heap/point-based-heap/field-insensitive/analysis___untyped_A.f6(A_this).dot
@@ -1,9 +1,9 @@
 digraph {
-	"node0" [shape="rect",color="black",label=<x = new int[](3)<BR/>{{<BR/>heap [[ [heap[s]:pp@'imp-testcases/heap/point-based-heap/field-insensitive/program.imp':54:19] ]]<BR/>value [[ #TOP# ]]<BR/>}} -&gt; [x]>];
+	"node0" [shape="rect",color="black",label=<x = new int[](3)<BR/>{{<BR/>heap [[ [heap[w]:pp@'imp-testcases/heap/point-based-heap/field-insensitive/program.imp':54:19] ]]<BR/>value [[ #TOP# ]]<BR/>}} -&gt; [x]>];
 	"node1" [shape="rect",color="gray",label=<a = +([](x, 0), [](y, 0))<BR/>{{<BR/>heap [[ [heap[w]:pp@'imp-testcases/heap/point-based-heap/field-insensitive/program.imp':54:19] ]]<BR/>value [[ a: [2, 2]<BR/>heap[w]:pp@'imp-testcases/heap/point-based-heap/field-insensitive/program.imp':54:19: [1, 1] ]]<BR/>}} -&gt; [a]>];
 	"node2" [shape="rect",color="gray",label=<[](x, 0) = 1<BR/>{{<BR/>heap [[ [heap[w]:pp@'imp-testcases/heap/point-based-heap/field-insensitive/program.imp':54:19] ]]<BR/>value [[ heap[w]:pp@'imp-testcases/heap/point-based-heap/field-insensitive/program.imp':54:19: [1, 1] ]]<BR/>}} -&gt; [heap[w]:pp@'imp-testcases/heap/point-based-heap/field-insensitive/program.imp':54:19]>];
 	"node3" [shape="rect",color="black",peripheries="2",label=<ret<BR/>{{<BR/>heap [[ [heap[w]:pp@'imp-testcases/heap/point-based-heap/field-insensitive/program.imp':54:19] ]]<BR/>value [[ a: [2, 2]<BR/>heap[w]:pp@'imp-testcases/heap/point-based-heap/field-insensitive/program.imp':54:19: [1, 1] ]]<BR/>}} -&gt; [skip]>];
-	"node4" [shape="rect",color="gray",label=<y = x<BR/>{{<BR/>heap [[ [heap[s]:pp@'imp-testcases/heap/point-based-heap/field-insensitive/program.imp':54:19] ]]<BR/>value [[ #TOP# ]]<BR/>}} -&gt; [y]>];
+	"node4" [shape="rect",color="gray",label=<y = x<BR/>{{<BR/>heap [[ [heap[w]:pp@'imp-testcases/heap/point-based-heap/field-insensitive/program.imp':54:19] ]]<BR/>value [[ #TOP# ]]<BR/>}} -&gt; [y]>];
 	"node1" -> "node3" [color="black"];
 	"node2" -> "node1" [color="black"];
 	"node0" -> "node4" [color="black"];
