+++ lisa/imp-testcases/heap/point-based-heap/field-insensitive/analysis___untyped_A.f7(A_this).dot
@@ -1,10 +1,10 @@
 digraph {
-	"node0" [shape="rect",color="black",label=<x = new int[](3)<BR/>{{<BR/>heap [[ [heap[s]:pp@'imp-testcases/heap/point-based-heap/field-insensitive/program.imp':61:19] ]]<BR/>value [[ #TOP# ]]<BR/>}} -&gt; [x]>];
-	"node1" [shape="rect",color="gray",label=<z = y<BR/>{{<BR/>heap [[ [heap[s]:pp@'imp-testcases/heap/point-based-heap/field-insensitive/program.imp':61:19, heap[s]:pp@'imp-testcases/heap/point-based-heap/field-insensitive/program.imp':62:19] ]]<BR/>value [[ #TOP# ]]<BR/>}} -&gt; [z]>];
-	"node2" [shape="rect",color="gray",label=<[](z, 0) = 1<BR/>{{<BR/>heap [[ [heap[s]:pp@'imp-testcases/heap/point-based-heap/field-insensitive/program.imp':61:19, heap[w]:pp@'imp-testcases/heap/point-based-heap/field-insensitive/program.imp':62:19] ]]<BR/>value [[ heap[w]:pp@'imp-testcases/heap/point-based-heap/field-insensitive/program.imp':62:19: [1, 1] ]]<BR/>}} -&gt; [heap[w]:pp@'imp-testcases/heap/point-based-heap/field-insensitive/program.imp':62:19]>];
-	"node3" [shape="rect",color="black",peripheries="2",label=<ret<BR/>{{<BR/>heap [[ [heap[s]:pp@'imp-testcases/heap/point-based-heap/field-insensitive/program.imp':61:19, heap[w]:pp@'imp-testcases/heap/point-based-heap/field-insensitive/program.imp':62:19] ]]<BR/>value [[ heap[w]:pp@'imp-testcases/heap/point-based-heap/field-insensitive/program.imp':62:19: [1, 1] ]]<BR/>}} -&gt; [skip]>];
-	"node4" [shape="rect",color="gray",label=<y = new int[](3)<BR/>{{<BR/>heap [[ [heap[s]:pp@'imp-testcases/heap/point-based-heap/field-insensitive/program.imp':61:19, heap[s]:pp@'imp-testcases/heap/point-based-heap/field-insensitive/program.imp':62:19] ]]<BR/>value [[ #TOP# ]]<BR/>}} -&gt; [y]>];
-	"node5" [shape="rect",color="gray",label=<z = x<BR/>{{<BR/>heap [[ [heap[s]:pp@'imp-testcases/heap/point-based-heap/field-insensitive/program.imp':61:19, heap[s]:pp@'imp-testcases/heap/point-based-heap/field-insensitive/program.imp':62:19] ]]<BR/>value [[ #TOP# ]]<BR/>}} -&gt; [z]>];
+	"node0" [shape="rect",color="black",label=<x = new int[](3)<BR/>{{<BR/>heap [[ [heap[w]:pp@'imp-testcases/heap/point-based-heap/field-insensitive/program.imp':61:19] ]]<BR/>value [[ #TOP# ]]<BR/>}} -&gt; [x]>];
+	"node1" [shape="rect",color="gray",label=<z = y<BR/>{{<BR/>heap [[ [heap[w]:pp@'imp-testcases/heap/point-based-heap/field-insensitive/program.imp':61:19, heap[w]:pp@'imp-testcases/heap/point-based-heap/field-insensitive/program.imp':62:19] ]]<BR/>value [[ #TOP# ]]<BR/>}} -&gt; [z]>];
+	"node2" [shape="rect",color="gray",label=<[](z, 0) = 1<BR/>{{<BR/>heap [[ [heap[w]:pp@'imp-testcases/heap/point-based-heap/field-insensitive/program.imp':61:19, heap[w]:pp@'imp-testcases/heap/point-based-heap/field-insensitive/program.imp':62:19] ]]<BR/>value [[ heap[w]:pp@'imp-testcases/heap/point-based-heap/field-insensitive/program.imp':62:19: [1, 1] ]]<BR/>}} -&gt; [heap[w]:pp@'imp-testcases/heap/point-based-heap/field-insensitive/program.imp':62:19]>];
+	"node3" [shape="rect",color="black",peripheries="2",label=<ret<BR/>{{<BR/>heap [[ [heap[w]:pp@'imp-testcases/heap/point-based-heap/field-insensitive/program.imp':61:19, heap[w]:pp@'imp-testcases/heap/point-based-heap/field-insensitive/program.imp':62:19] ]]<BR/>value [[ heap[w]:pp@'imp-testcases/heap/point-based-heap/field-insensitive/program.imp':62:19: [1, 1] ]]<BR/>}} -&gt; [skip]>];
+	"node4" [shape="rect",color="gray",label=<y = new int[](3)<BR/>{{<BR/>heap [[ [heap[w]:pp@'imp-testcases/heap/point-based-heap/field-insensitive/program.imp':61:19, heap[w]:pp@'imp-testcases/heap/point-based-heap/field-insensitive/program.imp':62:19] ]]<BR/>value [[ #TOP# ]]<BR/>}} -&gt; [y]>];
+	"node5" [shape="rect",color="gray",label=<z = x<BR/>{{<BR/>heap [[ [heap[w]:pp@'imp-testcases/heap/point-based-heap/field-insensitive/program.imp':61:19, heap[w]:pp@'imp-testcases/heap/point-based-heap/field-insensitive/program.imp':62:19] ]]<BR/>value [[ #TOP# ]]<BR/>}} -&gt; [z]>];
 	"node1" -> "node2" [color="black"];
 	"node2" -> "node3" [color="black"];
 	"node4" -> "node5" [color="black"];
