+++ lisa/src/main/java/it/unive/lisa/analysis/impl/heap/pointbased/PointBasedHeap.java
@@ -1,11 +1,8 @@
 package it.unive.lisa.analysis.impl.heap.pointbased;
 
 import java.util.Collections;
-import java.util.HashMap;
 import java.util.HashSet;
 import java.util.List;
-import java.util.Map;
-import java.util.Map.Entry;
 import java.util.Set;
 
 import it.unive.lisa.analysis.Lattice;
@@ -45,11 +42,6 @@ import it.unive.lisa.symbolic.value.ValueExpression;
 public class PointBasedHeap extends BaseHeapDomain<PointBasedHeap> {
 
 	/**
-	 * The list of heap replacement
-	 */
-	protected final AllocationSites toWeaken;
-
-	/**
 	 * An heap environment tracking which allocation sites are associated to
 	 * each identifier.
 	 */
@@ -59,7 +51,7 @@ public class PointBasedHeap extends BaseHeapDomain<PointBasedHeap> {
 	 * Builds a new instance of field-insensitive point-based heap.
 	 */
 	public PointBasedHeap() {
-		this(new HeapEnvironment<AllocationSites>(new AllocationSites()), new AllocationSites());
+		this(new HeapEnvironment<AllocationSites>(new AllocationSites()));
 	}
 
 	/**
@@ -69,9 +61,8 @@ public class PointBasedHeap extends BaseHeapDomain<PointBasedHeap> {
 	 * @param heapEnv       the heap environment that this instance tracks
 	 * @param substitutions the list of heap replacement
 	 */
-	protected PointBasedHeap(HeapEnvironment<AllocationSites> heapEnv, AllocationSites toWeaken) {
+	protected PointBasedHeap(HeapEnvironment<AllocationSites> heapEnv) {
 		this.heapEnv = heapEnv;
-		this.toWeaken = toWeaken;
 	}
 
 	/**
@@ -98,17 +89,9 @@ public class PointBasedHeap extends BaseHeapDomain<PointBasedHeap> {
 				MemoryPointer pid = (MemoryPointer) exp;	
 				Identifier v = id instanceof MemoryPointer ? ((MemoryPointer) id).getLocation() : id;
 				HeapEnvironment<AllocationSites> heap = sss.heapEnv.assign(v, pid.getLocation(), pp);
-				result = result.lub(from(new PointBasedHeap(weaken(heap, sss.toWeaken), sss.toWeaken)));
+				result = result.lub(from(new PointBasedHeap(heap)));
 			} else
 				result = result.lub(sss);
-		}
-
-
-		if (id instanceof AllocationSite) {
-			Set<AllocationSite> subs = new HashSet<>(result.toWeaken.elements());
-			subs.add(new AllocationSite(id.getTypes(), ((AllocationSite) id).getId(), id.isWeak()));
-			AllocationSites newToWeaken = toWeaken.mk(subs);
-			result = from(new PointBasedHeap(weaken(result.heapEnv, newToWeaken), newToWeaken));
 		}
 
 		return result;
@@ -122,7 +105,7 @@ public class PointBasedHeap extends BaseHeapDomain<PointBasedHeap> {
 
 	@Override
 	public PointBasedHeap forgetIdentifier(Identifier id) throws SemanticException {
-		return from(new PointBasedHeap(heapEnv.forgetIdentifier(id), toWeaken));
+		return from(new PointBasedHeap(heapEnv.forgetIdentifier(id)));
 	}
 
 	@Override
@@ -149,7 +132,7 @@ public class PointBasedHeap extends BaseHeapDomain<PointBasedHeap> {
 
 	@Override
 	public PointBasedHeap top() {
-		return from(new PointBasedHeap(heapEnv.top(), toWeaken.top()));
+		return from(new PointBasedHeap(heapEnv.top()));
 	}
 
 	@Override
@@ -159,7 +142,7 @@ public class PointBasedHeap extends BaseHeapDomain<PointBasedHeap> {
 
 	@Override
 	public PointBasedHeap bottom() {
-		return from(new PointBasedHeap(heapEnv.bottom(), toWeaken.bottom()));
+		return from(new PointBasedHeap(heapEnv.bottom()));
 	}
 
 	@Override
@@ -174,12 +157,12 @@ public class PointBasedHeap extends BaseHeapDomain<PointBasedHeap> {
 
 	@Override
 	public PointBasedHeap mk(PointBasedHeap reference) {
-		return from(new PointBasedHeap(reference.heapEnv, reference.toWeaken));
+		return from(new PointBasedHeap(reference.heapEnv));
 	}
 
 	@Override
 	protected PointBasedHeap lubAux(PointBasedHeap other) throws SemanticException {
-		return from(new PointBasedHeap(heapEnv.lub(other.heapEnv), toWeaken.lub(other.toWeaken)));
+		return from(new PointBasedHeap(heapEnv.lub(other.heapEnv)));
 	}
 
 	@Override
@@ -197,7 +180,6 @@ public class PointBasedHeap extends BaseHeapDomain<PointBasedHeap> {
 		final int prime = 31;
 		int result = 1;
 		result = prime * result + ((heapEnv == null) ? 0 : heapEnv.hashCode());
-		result = prime * result + ((toWeaken == null) ? 0 : toWeaken.hashCode());
 		return result;
 	}
 
@@ -215,51 +197,7 @@ public class PointBasedHeap extends BaseHeapDomain<PointBasedHeap> {
 				return false;
 		} else if (!heapEnv.equals(other.heapEnv))
 			return false;
-		if (toWeaken == null) {
-			if (other.toWeaken != null)
-				return false;
-		} else if (!toWeaken.equals(other.toWeaken))
-			return false;
 		return true;
-	}
-
-	private AllocationSite alreadyAllocated(AllocationSite id) {
-		for (AllocationSites set : heapEnv.values())
-			for (AllocationSite site : set)
-				if (site.getName().equals(id.getName()))
-					return site;
-
-		return null;
-	}
-
-	/**
-	 * Applies a substitution of identifiers substitutions in a given heap
-	 * environment.
-	 * 
-	 * @param heap         the heap environment where to apply the substitutions
-	 * @param toWeaken the substitution to apply
-	 * 
-	 * @return the heap environment modified by the substitution
-	 */
-	protected HeapEnvironment<AllocationSites> weaken(HeapEnvironment<AllocationSites> heap,
-			AllocationSites toWeaken) {
-		if (heap.isTop() || heap.isBottom() || toWeaken.isEmpty())
-			return heap;
-
-		Map<Identifier, AllocationSites> map = new HashMap<>();
-
-		for (Entry<Identifier, AllocationSites> entry : heap) {
-			Set<AllocationSite> newSites = new HashSet<>();
-			for (AllocationSite l : entry.getValue())
-				if (toWeaken.elements().stream().noneMatch(t -> t.getId().equals(l.getId())))
-					newSites.add(l);
-				else
-					newSites.add(new AllocationSite(l.getTypes(), l.getId(), true));
-
-			map.put(entry.getKey(), new AllocationSites().mk(newSites));
-		}
-
-		return new HeapEnvironment<>(new AllocationSites(), map);
 	}
 
 	@Override
@@ -271,7 +209,7 @@ public class PointBasedHeap extends BaseHeapDomain<PointBasedHeap> {
 		}
 
 		if (expression instanceof HeapAllocation)
-			return from(new PointBasedHeap(heapEnv, toWeaken));
+			return this;
 
 		if (expression instanceof HeapReference)
 			return smallStepSemantics(((HeapReference) expression).getExpression(), pp);
@@ -279,7 +217,6 @@ public class PointBasedHeap extends BaseHeapDomain<PointBasedHeap> {
 		if (expression instanceof HeapDereference) 
 			return smallStepSemantics(((HeapDereference) expression).getExpression(), pp);
 			
-
 		return top();
 	}
 
@@ -299,12 +236,12 @@ public class PointBasedHeap extends BaseHeapDomain<PointBasedHeap> {
 
 	@Override
 	public PointBasedHeap popScope(ScopeToken scope) throws SemanticException {
-		return new PointBasedHeap(heapEnv.popScope(scope), toWeaken);
+		return new PointBasedHeap(heapEnv.popScope(scope));
 	}
 
 	@Override
 	public PointBasedHeap pushScope(ScopeToken scope) throws SemanticException {
-		return new PointBasedHeap(heapEnv.pushScope(scope), toWeaken);
+		return new PointBasedHeap(heapEnv.pushScope(scope));
 	}
 
 	/**
@@ -335,11 +272,7 @@ public class PointBasedHeap extends BaseHeapDomain<PointBasedHeap> {
 				throws SemanticException {
 			AllocationSite id = new AllocationSite(expression.getTypes(),
 					((ProgramPoint) params[0]).getLocation().getCodeLocation());
-
-			if (alreadyAllocated(id) != null) 
-				return new ExpressionSet<>(new AllocationSite(id.getTypes(), id.getId(), true));
-			else
-				return new ExpressionSet<>(new AllocationSite(id.getTypes(), id.getId(), false));
+			return new ExpressionSet<>(new AllocationSite(id.getTypes(), id.getId(), true));
 		}
 
 		@Override
